fix issue with paint loading logic, fix issues related to recent appbar css changes
