+++ src/client/features/pathways/index.js
@@ -13,7 +13,6 @@ const Sidebar = require('../../common/components/sidebar');
 const EmptyNetwork = require('../../common/components/empty-network');
 const InfoMenu = require('./menus/network-info-menu');
 const FileDownloadMenu = require('./menus/file-download-menu');
-const datasourceToUrl = require('../../common/datasource-url-map');
 
 const Pathway = require('../../models/pathway/pathway-model');
 
@@ -88,8 +87,6 @@ class Pathways extends React.Component {
   render() {
     let { loading, pathway, cySrv, activeMenu, networkEmpty } = this.state;
 
-    const databaseURL = datasourceToUrl(pathway.datasource());
-
     let network = h('div.network', { className: classNames({
       'network-loading': loading,
       'network-sidebar-open': activeMenu !== 'closeMenu'
@@ -107,14 +104,12 @@ class Pathways extends React.Component {
       h('div.app-bar-branding', [
         h('i.app-bar-logo', { href: 'http://www.pathwaycommons.org/' }),
         h('div.app-bar-title', [
-          h('a.plain-link', { href:databaseURL,target:'_blank' }, ' ' + pathway.datasource()),
+          h('a.plain-link', { href: pathway.datasourceUrl(), target: '_blank' }, ' ' + pathway.datasource()),
           h('span', ': ' + pathway.name())
         ])
       ]),
       toolbar
     ]);
-
-
 
     let sidebar = h('div.app-sidebar', [
       h(Sidebar, {  controller: this, activeMenu }, [
