+++ src/client/models/pathway/pathway-model.js
@@ -1,4 +1,17 @@
 const _ = require('lodash');
+const diceCoefficient = require('dice-coefficient');
+
+let datasourceURLMap = new Map();
+
+datasourceURLMap.set('humancyc','https://humancyc.org'); // HumanCyc
+datasourceURLMap.set('inoh','https://www.ncbi.nlm.nih.gov/pubmed/22120663'); //Integrating Network Objects with Hierarchies
+datasourceURLMap.set('kegg','https://www.kegg.jp'); // KEGG Pathway
+datasourceURLMap.set('pid','https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2686461/'); // NCI Pathway Interaction Database: Pathway
+datasourceURLMap.set('netpath','http://www.netpath.org'); // NetPath
+datasourceURLMap.set('panther','http://www.pantherdb.org'); // PANTHER Pathway
+datasourceURLMap.set('reactome','https://reactome.org'); // Reactome
+datasourceURLMap.set('smpdb','http://smpdb.ca'); // Small Molecule pathway Database
+datasourceURLMap.set('wikipathways','https://www.wikipathways.org'); // WikiPathways
 
 // basic pathway model to serve the needs of the pathways view
 // TODO:
@@ -33,6 +46,23 @@ class Pathway {
 
   datasource(){
     return _.get(this.raw, 'graph.pathwayMetadata.dataSource.0', 'Unknown datasource');
+  }
+
+  datasourceUrl(){
+    let datasource = this.datasource();
+    let lowercaseName = datasource.toLowerCase();
+
+    let databaseURL = datasourceURLMap.get(lowercaseName);
+    if( !databaseURL ){
+      let datasourceURLScores = [];
+      datasourceURLMap.forEach( (datasourceURL, datasource) => {
+        datasourceURLScores.push([datasourceURL,diceCoefficient(datasource,lowercaseName)]);
+      });
+      datasourceURLScores.sort( (a,b) => { return b[1] - a[1]; } );
+      databaseURL = datasourceURLScores[0][0];
+    }
+
+    return databaseURL;
   }
 
   comments(){
