+++ src/styles/features/pathways/app-bar.css
@@ -33,11 +33,12 @@
 .app-toolbar {
   z-index: 10;
   position: relative;
-  top:0px;
-  width:350px;
+  top: 0px;
+  width: 350px;
   border-radius: 0 0 0.25em 0.25em;
   border-color: var(--light-base-colour-dark);
   border-style: solid;
   border-width: 1px;
   background-color:white;
+  opacity: 0.9;
 }
